Refactor: clean up accelerator
