+++ src/accelerate/utils/__init__.py
@@ -220,6 +220,7 @@ from .fsdp_utils import (
     enable_fsdp_ram_efficient_loading,
     ensure_weights_retied,
     fsdp2_load_full_state_dict,
+    fsdp2_prepare_model,
     fsdp2_switch_optimizer_parameters,
     load_fsdp_model,
     load_fsdp_optimizer,
@@ -267,6 +268,7 @@ from .other import (
     clean_state_dict_for_safetensors,
     convert_bytes,
     extract_model_from_parallel,
+    get_children_bottom_up,
     get_pretty_name,
     is_port_in_use,
     load,
